lorax: Skip rpm database and /boot removal
The standard lorax templates remove the rpm database and most of the
/boot contents. Because of how osbuild runs the scripts we need to keep
the database and /boot around until later.

So this skips removing the rpm database, and skips any command operating
on /boot.

These files can be excluded from the install.img by adding them to the
squashfs/erofs stage's exclude_paths list.

With these changes it is now possible to add a cleanup stage, after the
dracut stage inside the anaconda-tree pipeline, which will remove the
same files that are removed on the Lorax produced boot.iso

eg. add this after the org.osbuild.dracut stage:

        {
          "type": "org.osbuild.lorax-script",
          "options": {
            "path": "99-generic/runtime-cleanup.tmpl",
            "basearch": "x86_64",
            "product": {
              "name": "",
              "version": ""
            },
            "branding": {
              "release": "fedora-release",
              "logos": "fedora-logos"
            }
          }
        },

And add the excluded files/paths to the org.osbuild.squashfs stage:

          "type": "org.osbuild.squashfs",
          "inputs": {
            "tree": {
              "type": "org.osbuild.tree",
              "origin": "org.osbuild.pipeline",
              "references": [
                "name:anaconda-tree"
              ]
            }
          },
          "options": {
            "filename": "images/install.img",
            "exclude_paths": [
              "boot/efi/.*",
              "boot/grub2/.*",
              "boot/config-.*",
              "boot/initramfs-.*",
              "boot/loader/.*",
              "boot/symvers-.*",
              "boot/System.map-.*",
              "usr/lib/sysimage/rpm/.*",
              "var/lib/rpm/.*",
              "var/lib/yum/.*",
              "var/lib/dnf/.*"
            ],
            "compression": {
              "method": "xz",
              "options": {
                "bcj": "x86"
              }
            }
          }
        },

+++ osbuild/util/lorax.py
@@ -155,7 +155,14 @@ class Script:
 
     @command
     def remove(self, *files):
+        skiplist = ["/usr/lib/sysimage/rpm/", "/var/lib/rpm/", "/var/lib/yum", "/var/lib/dnf"]
         for g in files:
+            # We need the rpm database in order to remove files from packages, so skip removing
+            # it. This can be handled by skipping these paths in the squashfs/erofs stage.
+            if any(g.startswith(p) for p in skiplist):
+                print(f"remove: Skipping RPM database path {g}")
+                continue
+
             for f in rglob(self.tree_path(g)):
                 if os.path.isdir(f) and not os.path.islink(f):
                     shutil.rmtree(f)
@@ -177,6 +184,12 @@ class Script:
 
     @command
     def runcmd(self, *args):
+        # We need /boot content in order to create the iso, so skip the commands acting on
+        # boot. This can be handled by skipping these paths in the squashfs/erofs stage.
+        if any("/boot" in a for a in args):
+            print("runcmd: skipping operations on /boot")
+            return
+
         print("run ", " ".join(args))
         subprocess.run(args, cwd=self.tree, check=True)
